Add ChangeUserForm component and integrate user editing functionality
